italian words + readme updates
