+++ packages/english-verbs-gerunds/example.js
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 const EnglishGerunds = require('./dist/gerunds.json');
 
 // referring
