+++ packages/english-verbs-helper/README.md
@@ -50,8 +50,8 @@ npm install english-verbs-helper
 
 ```javascript
 const EnglishVerbs = require('english-verbs-helper');
-const Irregular = require('english-verbs-irregular');
-const Gerunds = require('english-verbs-gerunds');
+const Irregular = require('english-verbs-irregular/dist/verbs.json');
+const Gerunds = require('english-verbs-gerunds/dist/gerunds.json');
 
 const VerbsData = EnglishVerbs.mergeVerbsData(Irregular, Gerunds);
