+++ packages/german-verbs/README.md
@@ -17,7 +17,7 @@ npm install german-verbs
 
 ```javascript
 const GermanVerbsLib = require('german-verbs');
-const GermanVerbsDict = require('german-verbs-dict');
+const GermanVerbsDict = require('german-verbs-dict/dist/verbs.json');
 
 // hörten
 console.log(GermanVerbsLib.getConjugation(GermanVerbsDict, 'hören', 'PRATERITUM', 3, 'P'));
