+++ packages/italian-words-dict/test/test.js
@@ -8,13 +8,14 @@ const assert = require('assert');
 const italianWords = require('../dist/words.json');
 
 describe('italian-words-dict', function () {
-  it('should work', function (done) {
+  it('should contain something', function () {
     assert(italianWords != null);
     assert(Object.keys(italianWords).length > 100);
+  });
+  it('pizza should be ok', function () {
     const pizza = italianWords['pizza'];
     assert(pizza != null);
     assert.strictEqual(pizza['G'], 'F');
     assert.strictEqual(pizza['P'], 'pizze');
-    done();
   });
 });
