+++ packages/italian-words-dict/tsconfig.json
@@ -6,7 +6,7 @@
       "../../node_modules/@types"
     ],
     "sourceMap": true,
-    "outDir": "dist/create",
+    "outDir": "dist",
     "declaration": true,
     "resolveJsonModule": true,
     "esModuleInterop": true
