+++ packages/rosaenlg/src/LanguageItalian.ts
@@ -12,8 +12,9 @@ import { VerbsData } from 'rosaenlg-pug-code-gen';
 import { getDet as getItalianDet, DetType as ItalianDetType, Dist as ItalianDist } from 'italian-determiners';
 import { agreeItalianAdjective } from 'italian-adjectives';
 import italianAdjectivesDict from 'italian-adjectives-dict/dist/adjectives.json';
-import { getGenderItalianWord, getNumberItalianWord, WordsInfo as ItalianWordsInfo } from 'italian-words';
-import italianWordsDict from 'italian-words-dict';
+import { getGenderItalianWord, getNumberItalianWord } from 'italian-words';
+import { WordsInfo as ItalianWordsInfo } from 'italian-words-dict';
+import italianWordsDict from 'italian-words-dict/dist/words.json';
 import { getOrdinal as getItalianOrdinal } from 'italian-ordinals-cardinals';
 import 'numeral/locales/it';
 import { it as dataFnsIt } from 'date-fns/locale';
